+++ torchtune/modules/__init__.py
@@ -4,12 +4,16 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-from .attention import CausalSelfAttention  # noqa
+from .attention import CausalSelfAttentio, GroupedQueryAttention  # noqa
 from .common_utils import reparametrize_as_dtype_state_dict_post_hook
 from .feed_forward import FeedForward  # noqa
 from .kv_cache import KVCache  # noqa
 from .layer_norm import Fp32LayerNorm  # noqa
 from .lr_schedulers import get_cosine_schedule_with_warmup  # noqa
+from .multimodal_transformer import (  # noqa
+    TransformerCrossAttentionLayer,
+    TransformerSelfAttentionLayer,
+)
 from .position_embeddings import RotaryPositionalEmbeddings  # noqa
 from .rms_norm import RMSNorm  # noqa
 from .transformer import TransformerDecoder, TransformerDecoderLayer  # noqa
@@ -17,6 +21,7 @@ from .vision_transformer import VisionTransformer
 
 __all__ = [
     "CausalSelfAttention",
+    "GroupedQueryAttention",
     "FeedForward",
     "get_cosine_schedule_with_warmup",
     "KVCache",
@@ -26,5 +31,7 @@ __all__ = [
     "VisionTransformer",
     "TransformerDecoder",
     "TransformerDecoderLayer",
+    "TransformerSelfAttentionLayer",
+    "TransformerCrossAttentionLayer",
     "reparametrize_as_dtype_state_dict_post_hook",
 ]
